Refactor: create eventstore package
