+++ storage/Storage.go
@@ -1,0 +1,10 @@
+package storage
+
+type Document struct {
+	ID   string
+	Body interface{}
+}
+
+type Storage interface {
+	Save(index string, document *Document) error
+}
