+++ storage/eventstore/Event.go
@@ -1,4 +1,4 @@
-package storage
+package eventstore
 
 import (
 	"time"
